Move ednpoint tests into a dedicated sub namespace
